Add Options panel and match up HUD and receiver scripts
The options panel goes opposite the skin panel on the HUD and will
contain the controls for fingernails and toenails.

I've gone ahead and just made two new scripts for this, there is
still some of Shin's code here but these should be considered new
for most purposes.

They both still use the same commands for alpha and skin and are
compatible with RC2 and RC# except for possible appID mismatches.