Fix receiver script ALPHA bug - Contrib Edition
See PR https://github.com/RuthAndRoth/Ruth/pull/7 for the details,
I am changing these after talking with Austin, the idea being that
this fixes a subtle bug and we do not want it propogated if the
Contrib scripts are copied.

We can revert this if folks do not want any external changes
in Contrib.