HUD: Break long messages into chunks in receiver
LSL has a limit of 1024 bytes tin llSay() family functions, the element
map for Ruth2 v3 is longer than that in JSON form.  Add a chunking
capability to send_csv() to break long messages into chunks.

Chunked messages have the same command (first word in the command list)
prepended with a '+' char so the receiver knows to buffer the messages.
The last chunk does not have the '+' so the receiver can handle it in
the same flow, checking to grab anything previously stored in the buffer.