very early draft on Ruth bento head
changing the way I approach this kind of modeling, this time starting with extreme lo-poly/LOD, working on setting it up so that higher poly versions will have the identical UV mapping, so we can use them for LOD uploads. Plus a very high poly version for baking normal maps.