ss HUD: Simplify alpha HUD mesh, add OSGrid UUIDs
The prior alpha HUD mesh was derived from Shin's original alpha HUD and
was a mesh linkset of 6 or more objects.  The buttons were removed from
that linkset to use the 'outline buttons' used elsewhere in this HUD,
leaving only two objects, one of which was the rotate bar at the bottom
with an unnecessary face configuration.

That mesh is now replaced with a simple mesh ('box_inside.dae') that
replicates the inside 4 walls of the box.  The rotatebar is now a simple
prim and rotation will be handled similarly to the menu bar.