reassemble, refit, remap
After making the topography match slim (using shrinkwrap), I resourced the hands from an earlier file so they'd fit and we'd have UV mapping for those who don't like wearing separate fingernails, aka "talons". Also re-mapped the UVs for the feet similar to Roth, and slimmed the ankles and feet.  Reshaped the shoulders so they'll weight more naturally and started fixing the weights, so we don't get that pinched look under the arms.