Moved contribution to its own Contrib/MyraFoil directory
This is a further development of Duck Girl's RuthRC3_BakesOnMesh_Slim 2.dae. To better fit clothing made for Maitreya the shoulders are lifted a bit. The groin is changed and the UV map is fixed at the groin to prevent to much stretching of the texture there.

Capitalisation of RuthToo RC3 HoB Upper v3.dae changed.